+++ src/cli/delete_image.rs
@@ -1,10 +1,11 @@
-use imgurs::api::{delete_image::delete_image as del_img, ImgurClient};
+use colored::Colorize;
 
-use log::info;
+use imgurs::api::{delete_image::delete_image as del_img, ImgurClient};
 
 pub async fn delete_image(client: ImgurClient, delete_hash: String) {
     let i = del_img(client, delete_hash)
         .await
         .expect("send api request");
-    info!("{i}");
+
+    println!("{}", i.magenta());
 }