+++ CHANGELOG.md
@@ -8,11 +8,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 <!-- next-header -->
 
 ## [Unreleased]
+
+## [0.4.0] - 2022-02-27
 ### CLI
 - update logger
 - added clipboard
 - added manpage
 - added completion for elvish
+- if failed to upload image send notify with error message
 
 ### Library
 - added Clone derive
@@ -60,7 +63,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - upload image
 
 <!-- next-url -->
-[Unreleased]: https://github.com/MedzikUser/imgurs/compare/v0.3.0...HEAD
+[Unreleased]: https://github.com/MedzikUser/imgurs/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/MedzikUser/imgurs/commits/v0.4.0
 [0.3.0]: https://github.com/MedzikUser/imgurs/commits/v0.3.0
 [0.2.0]: https://github.com/MedzikUser/imgurs/commits/v0.2.0
 [0.1.0]: https://github.com/MedzikUser/imgurs/commits/v0.1.0